+++ DNN Platform/Library/Services/FileSystem/FileManager.cs
@@ -284,7 +284,7 @@ namespace DotNetNuke.Services.FileSystem
                     }
 
                     // Retrieve Metadata
-                    file.Size = (int)fileContent.Length;
+                    file.Size = fileContent.Length;
                     file.SHA1Hash = GetHash(fileContent);
                     fileContent.Position = 0;
 
@@ -354,7 +354,7 @@ namespace DotNetNuke.Services.FileSystem
                 }
                 else
                 {
-                    file.Size = (int)folderProvider.GetFileSize(file);
+                    file.Size = folderProvider.GetFileSize(file);
                 }
 
                 if (folderWorkflow == null || !fileExists)
@@ -1063,7 +1063,7 @@ namespace DotNetNuke.Services.FileSystem
                 if (folderMapping != null)
                 {
                     var folderProvider = FolderProvider.Instance(folderMapping.FolderProviderType);
-                    file.Size = (int)folderProvider.GetFileSize(file);
+                    file.Size = folderProvider.GetFileSize(file);
                     file.LastModificationTime = folderProvider.GetLastModificationTime(file);
                 }
             }
